requirements.txt
python-twitter
yaml

